LPS-201309 Create SXPParameterContributorProvider, SXPParameterContributorProviderImpl, move sxpParameterContributors logic from SXPParameterDataCreator to SXPParameterContributorProviderImpl and use SXPParameterContributorProvider instead
