+++ modules/dxp/apps/search-experiences/search-experiences-api/src/main/java/com/liferay/search/experiences/blueprint/parameter/contributor/SXPParameterContributorProvider.java
@@ -1,0 +1,15 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.search.experiences.blueprint.parameter.contributor;
+
+/**
+ * @author Renan Vasconcelos
+ */
+public interface SXPParameterContributorProvider {
+
+	public SXPParameterContributor[] getSxpParameterContributors();
+
+}
