+++ modules/dxp/apps/search-experiences/search-experiences-service/src/main/java/com/liferay/search/experiences/internal/blueprint/parameter/SXPParameterContributorProviderImpl.java
@@ -1,0 +1,115 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.search.experiences.internal.blueprint.parameter;
+
+import com.liferay.asset.kernel.service.AssetCategoryLocalService;
+import com.liferay.asset.kernel.service.AssetTagLocalService;
+import com.liferay.expando.kernel.service.ExpandoColumnLocalService;
+import com.liferay.expando.kernel.service.ExpandoValueLocalService;
+import com.liferay.portal.configuration.module.configuration.ConfigurationProvider;
+import com.liferay.portal.kernel.language.Language;
+import com.liferay.portal.kernel.service.GroupLocalService;
+import com.liferay.portal.kernel.service.UserGroupGroupRoleLocalService;
+import com.liferay.portal.kernel.service.UserGroupLocalService;
+import com.liferay.portal.kernel.service.UserGroupRoleLocalService;
+import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.util.Portal;
+import com.liferay.search.experiences.blueprint.parameter.contributor.SXPParameterContributor;
+import com.liferay.search.experiences.blueprint.parameter.contributor.SXPParameterContributorProvider;
+import com.liferay.search.experiences.configuration.SemanticSearchConfigurationProvider;
+import com.liferay.search.experiences.internal.blueprint.parameter.contributor.ContextSXPParameterContributor;
+import com.liferay.search.experiences.internal.blueprint.parameter.contributor.IpstackSXPParameterContributor;
+import com.liferay.search.experiences.internal.blueprint.parameter.contributor.MLSXPParameterContributor;
+import com.liferay.search.experiences.internal.blueprint.parameter.contributor.OpenWeatherMapSXPParameterContributor;
+import com.liferay.search.experiences.internal.blueprint.parameter.contributor.TimeSXPParameterContributor;
+import com.liferay.search.experiences.internal.blueprint.parameter.contributor.UserSXPParameterContributor;
+import com.liferay.search.experiences.ml.embedding.text.TextEmbeddingRetriever;
+import com.liferay.segments.SegmentsEntryRetriever;
+
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Renan Vasconcelos
+ */
+@Component(enabled = false, service = SXPParameterContributorProvider.class)
+public class SXPParameterContributorProviderImpl
+	implements SXPParameterContributorProvider {
+
+	@Override
+	public SXPParameterContributor[] getSxpParameterContributors() {
+		return _sxpParameterContributors;
+	}
+
+	@Activate
+	protected void activate() {
+		_sxpParameterContributors = new SXPParameterContributor[] {
+			new ContextSXPParameterContributor(_groupLocalService, _language),
+			new IpstackSXPParameterContributor(_configurationProvider),
+			new MLSXPParameterContributor(
+				_language, _semanticSearchConfigurationProvider,
+				_textEmbeddingRetriever),
+			new OpenWeatherMapSXPParameterContributor(_configurationProvider),
+			new TimeSXPParameterContributor(),
+			new UserSXPParameterContributor(
+				_assetCategoryLocalService, _assetTagLocalService,
+				_expandoColumnLocalService, _expandoValueLocalService,
+				_groupLocalService, _language, _portal, _segmentsEntryRetriever,
+				_userGroupGroupRoleLocalService, _userGroupLocalService,
+				_userGroupRoleLocalService, _userLocalService)
+		};
+	}
+
+	@Reference
+	private AssetCategoryLocalService _assetCategoryLocalService;
+
+	@Reference
+	private AssetTagLocalService _assetTagLocalService;
+
+	@Reference
+	private ConfigurationProvider _configurationProvider;
+
+	@Reference
+	private ExpandoColumnLocalService _expandoColumnLocalService;
+
+	@Reference
+	private ExpandoValueLocalService _expandoValueLocalService;
+
+	@Reference
+	private GroupLocalService _groupLocalService;
+
+	@Reference
+	private Language _language;
+
+	@Reference
+	private Portal _portal;
+
+	@Reference
+	private SegmentsEntryRetriever _segmentsEntryRetriever;
+
+	@Reference
+	private SemanticSearchConfigurationProvider
+		_semanticSearchConfigurationProvider;
+
+	private SXPParameterContributor[] _sxpParameterContributors;
+
+	@Reference
+	private TextEmbeddingRetriever _textEmbeddingRetriever;
+
+	@Reference
+	private UserGroupGroupRoleLocalService _userGroupGroupRoleLocalService;
+
+	@Reference
+	private UserGroupLocalService _userGroupLocalService;
+
+	@Reference
+	private UserGroupRoleLocalService _userGroupRoleLocalService;
+
+	@Reference
+	private UserLocalService _userLocalService;
+
+}
